+++ react/utils/messages.ts
@@ -145,4 +145,9 @@ export const messages = defineMessages({
   importFileSubtitleModel: { id: 'admin/import.file.subtitle.model.text' },
   importFileSubtitle2: { id: 'admin/import.file.subtitle.2.text' },
   detailsLabel: { id: 'admin/details.label' },
+  orderStatusLabel: { id: 'admin/order.status.label' },
+  orderStatusCreatedLabel: { id: 'admin/order.status.created.label' },
+  orderStatusPaidLabel: { id: 'admin/order.status.paid.label' },
+  orderStatusCancelLabel: { id: 'admin/order.status.cancel.label' },
+  orderStatusInvoicedLabel: { id: 'admin/order.status.invoiced.label' },
 })